fitsync-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Core ML Libraries
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
mediapipe==0.10.8
transformers==4.35.2

# Advanced ML
tensorflow==2.15.0
diffusers==0.24.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
insightface==0.7.3

# Database & Caching
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# Vector Database
chromadb==0.4.18
faiss-cpu==1.7.4

# Cloud & Storage
boto3==1.34.0
python-dotenv==1.0.0

# API & Validation
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional Dependencies for Enhanced Features
requests==2.31.0
aiofiles==23.2.1
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0